feat(preferences): migrate to Adw.PreferencesDialog
For this, the name source dialog is integrated as a sub page and the
illuminant settings have been removed.
