--- data/resources/ui/preferences/name-sources-dialog.blp
@@ -1,48 +1,0 @@
-using Gtk 4.0;
-using Adw 1;
-
-template $NameSourceDialog : Adw.Window {
-  title: _("Name Sources");
-  modal: true;
-  default-width: 450;
-  content: $AdwToolbarView {
-    [top]
-    Adw.HeaderBar {}
-    content: Box {
-      orientation: vertical;
-
-      Adw.PreferencesPage {
-        Adw.PreferencesGroup {
-          $AdwSwitchRow name_source_basic {
-            title: C_("Name of the basic color keyword set from https://www.w3.org/TR/css-color-3/#html4", "Basic");
-            subtitle: _("Basic web colors");
-          }
-
-          $AdwSwitchRow name_source_extended {
-            title: C_("Name of the extended color keyword set from https://www.w3.org/TR/css-color-3/#svg-color", "Extended");
-            subtitle: _("X11 and SVG color values");
-          }
-
-          $AdwSwitchRow name_source_gnome {
-            title: C_("Name of the color set from the GNOME color palette (https://developer.gnome.org/hig/reference/palette.html)", "GNOME Color Palette");
-            subtitle: _("Colors for GNOME app icons and illustrations");
-          }
-
-          $AdwSwitchRow name_source_xkcd {
-            title: C_("Name of the color set from the xkcd color survey", "xkcd Color Survey");
-            subtitle: _("954 RGB colors named by volunteers");
-          }
-        }
-      }
-
-      ShortcutController {
-        scope: managed;
-
-        Shortcut {
-          trigger: "Escape|<Ctrl>w";
-          action: "action(window.close)";
-        }
-      }
-    };
-  };
-}
