--- data/resources/ui/preferences/preferences.blp
@@ -1,135 +1,0 @@
-using Gtk 4.0;
-using Adw 1;
-
-template $PreferencesWindow : Adw.PreferencesWindow {
-  default-width: 550;
-  default-height: 570;
-  can-navigate-back: true;
-
-  Adw.PreferencesPage {
-    icon-name: "applications-system-symbolic";
-    title: _("General");
-
-    Adw.PreferencesGroup {
-      title: _("Formatting");
-
-      Adw.ComboRow alpha_pos_box {
-        title: _("Alpha-Value-Position");
-        subtitle: _("Where the Alphavalue is positioned in the Hexstring");
-        model: StringList {
-          strings [
-            C_("Alphavalue is not shown", "None"),
-            C_("Alphavalue is shown at the end", "End"),
-            C_("Alphavalue is shown at the start", "Start"),
-          ]
-        };
-      }
-
-      Adw.ActionRow {
-        title: _("Name Sources");
-        subtitle: _("Sources for displaying color names");
-        activatable: true;
-        activated => $on_name_row_activated() swapped;
-
-        [suffix]
-        Image {
-          icon-name: "go-next-symbolic";
-        }
-      }
-
-      $AdwSpinRow  precision_row {
-        title: _("Precision");
-          adjustment: Adjustment {
-            value: 2;
-            lower: 0;
-            upper: 15;
-            step-increment: 1;
-          };
-
-          tooltip-text: _("Precision");
-          climb-rate: 1;
-          numeric: true;
-          valign: center;
-      }
-    }
-
-    Adw.PreferencesGroup {
-      title: _("CIE Color Space");
-
-      Adw.ActionRow {
-        title: _("CIE Illuminants");
-        subtitle: _("Lighting conditions for the CIE color space");
-        activatable-widget: cie_illuminants_box;
-
-        [suffix]
-        DropDown cie_illuminants_box {
-          valign: center;
-          model: StringList {
-            strings [
-              _("A (Incandescent/tungsten)"),
-              _("B (Old direct sunlight at noon)"),
-              _("C (Old daylight)"),
-              _("D50 (ICC profile PCS)"),
-              _("D55 (Mid-morning daylight)"),
-              _("D65 (Daylight, sRGB, Adobe-RGB)"),
-              _("D75 (North sky daylight)"),
-              _("E (Equal energy)"),
-              _("F1 (Daylight Fluorescent)"),
-              _("F2 (Cool fluorescent)"),
-              _("F3 (White Fluorescent)"),
-              _("F4 (Warm White Fluorescent)"),
-              _("F5 (Daylight Fluorescent)"),
-              _("F6 (Lite White Fluorescent)"),
-              _("F7 (Daylight fluorescent, D65 simulator)"),
-              _("F8 (Sylvania F40, D50 simulator)"),
-              _("F9 (Cool White Fluorescent)"),
-              _("F10 (Ultralume 50, Philips TL85)"),
-              _("F11 (Ultralume 40, Philips TL84)"),
-              _("F12 (Ultralume 30, Philips TL83)"),
-            ]
-          };
-
-          styles [
-            "flat",
-          ]
-
-        }
-      }
-
-      Adw.ComboRow standard_observer_box {
-        title: _("CIE Standard Observer");
-        subtitle: _("The CIE standard observer\'s field of view");
-        model: StringList {
-          strings [
-            "2°",
-            "10°",
-          ]
-        };
-      }
-    }
-
-    Adw.PreferencesGroup {
-      title: _("Color Formats");
-      description: _("Customize the visible formats and in which order they are displayed");
-
-      [header-suffix]
-      Button {
-        icon-name: "edit-clear-symbolic";
-        tooltip-text: C_("Tooltip of the button to reset the format order", "Reset Order");
-        clicked => $on_reset_pressed() swapped;
-
-        styles [
-          "flat",
-        ]
-      }
-
-      ListBox order_list {
-        selection-mode: none;
-
-        styles [
-          "boxed-list",
-        ]
-      }
-    }
-  }
-}
