+++ src/widgets/preferences/mod.rs
@@ -1,3 +1,2 @@
 pub mod color_format;
-mod name_set_dialog;
 pub mod preferences_window;
